+++ lib/patterns/object.js
@@ -20,7 +20,7 @@ var ObjectSchema = module.exports = Schema.patterns.ObjectSchema = Schema.extend
     })
   },
 
-  validatePropsFunction: function (intance, properties, validatedFields, validation) {
+  validatePropsFunction: function (intance, properties, validatedFields, collectAllErrors, validation) {
     var allErrors = [];
     for (var k in properties) {
       var check = properties[k];
@@ -33,37 +33,15 @@ var ObjectSchema = module.exports = Schema.patterns.ObjectSchema = Schema.extend
         }
       }
 
-      if (numberOfMatches < check.min && !check.value.validate(undefined))
+      if (numberOfMatches < check.min && !check.value.validate(undefined)) {
         allErrors.push(intance + " doesn't match " + check + " check");
+        if (!collectAllErrors)
+          return allErrors;
+      }
     }
     return allErrors;
   },
-  errors: function(instance) {
-    var self = this;
-
-    if (instance == null)
-      return ( instance + ' is not Object' );
-
-    var validatedFields = {};
-    for (var k in instance) {
-      validatedFields[k] = false;
-    }
-
-    var regexpErrors =  self.validatePropsFunction(instance, self.regexpProps, validatedFields, function (checkKey, instanceKey, check, value) {
-      return check.key.test(instanceKey) && check.value.validate(value);
-    });
-
-    var stringErrors = self.validatePropsFunction(instance, self.stringProps, validatedFields, function (checkKey, instanceKey, check, value) {
-      return checkKey === instanceKey && check.value.validate(value);
-    });
-
-    var missedPropsErrors = {};
-    for (var k in validatedFields) {
-      var result = validatedFields[k];
-      if (!result && !self.other.validate(instance[k])) {
-        missedPropsErrors[k] = "Property " + k + " doesn't match any of checks";
-      }
-    }
+  collectErrorsAsObject: function (regexpErrors, stringErrors, missedPropsErrors) {
     var allErrors = {};
     for (var k in regexpErrors) {
       if (regexpErrors.hasOwnProperty(k))
@@ -77,12 +55,53 @@ var ObjectSchema = module.exports = Schema.patterns.ObjectSchema = Schema.extend
       if (missedPropsErrors.hasOwnProperty([k]))
         allErrors[k] = missedPropsErrors[k];
     }
+    return allErrors;
+  },
+  errorsList: function(instance, collectAllErrors) {
+    var self = this;
 
+    if (instance == null)
+      return ( instance + ' is not Object' );
+
+    var validatedFields = {};
+    for (var k in instance) {
+      validatedFields[k] = false;
+    }
+
+    var regexpErrors =  self.validatePropsFunction(instance, self.regexpProps, validatedFields, collectAllErrors, function (checkKey, instanceKey, check, value) {
+      return check.key.test(instanceKey) && check.value.validate(value);
+    });
+    //if we should not collect all errors, generate error object right away
+    if (!collectAllErrors && regexpErrors.length > 0)
+      return this.collectErrorsAsObject(regexpErrors, [], []);
+
+
+    var stringErrors = self.validatePropsFunction(instance, self.stringProps, validatedFields, collectAllErrors, function (checkKey, instanceKey, check, value) {
+      return checkKey === instanceKey && check.value.validate(value);
+    });
+    //if we should not collect all errors, generate error object right away
+    if (!collectAllErrors && stringErrors.length > 0)
+      return this.collectErrorsAsObject([], stringErrors, []);
+
+    var missedPropsErrors = {};
+    for (var k in validatedFields) {
+      var result = validatedFields[k];
+      if (!result && !self.other.validate(instance[k])) {
+        missedPropsErrors[k] = "Property " + k + " doesn't match any of checks";
+        if (!collectAllErrors)
+          return missedPropsErrors;
+      }
+    }
+    var allErrors = this.collectErrorsAsObject(regexpErrors, stringErrors, missedPropsErrors);
     return Object.keys(allErrors).length == 0 ? false : allErrors;
+  },
+  errors: function(instance) {
+    var self = this;
+    return self.errorsList(instance, true)
   },
   validate: function(instance) {
     var self = this;
-    return self.errors(instance) ? false : true;
+    return self.errorsList(instance, false) ? false : true;
   },
 
   toJSON: Schema.session(function() {
